Implement heuristics
Change open_nodes to a heap, so the Nodes now have a __lt__ method.
Node.__lt__ simply compares the heuristic cost, so that it's possible
to get the next node in O(1) time.
Check if node is in closed_nodes is still present (twice) and uses
the node's __hash__ and __eq__ to find equivalents, reducing
unnecessary searches.
The node's cost is saved in the node itself, and is calculated by
calling the calculate_cost method.
Currently the class already supports A*'s g and h cost, but an
admissible heuristic hasn't been found yet.