Big ass update
Implement admissible heuristic into A* search.
In this script, the Nodes themselves save their goal, and so it
takes a lot of weight from the heuristic function (where we would
need to iterate over the different permutations of goal).
This method also helps because it gives the program a way to choose
the most promising permutation of goals. It possibly spends more memory
because each node has a new field.
The if statement 'if dest_node not in closed_nodes' was re-added, since
it boosted the time from 2000ms to 15000ms, and so it is completely worth
for big problems